pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-agency"
version = "0.1.0"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn==0.24.0", 
    "python-jose==3.3.0",
    "passlib==1.7.4",
    "python-multipart==0.0.6",
    "pydantic==2.4.2",
    "python-dotenv==1.0.0",
    "supabase==2.0.3",
    "openai==1.3.0",
    "pydantic-settings==2.0.3"
]